pyproject.toml
[project]
name = "cammana"
version = "2.0.0"
description = "ONVIF Camera Manager - Desktop Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.0,<2.4.0",
    "onvif-zeep>=0.2.12",
    "opencv-python>=4.11.0.86",
    "pillow>=12.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "PySide6>=6.6.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "ultralytics>=8.3.250",
    "httpx>=0.28.1",
    "apscheduler>=3.11.2",
    "fpdf2>=2.8.5",
    "python-jose>=3.5.0",
    "bcrypt==3.2.0",
    "passlib>=1.7.4",
    "zeroconf>=0.148.0",
    "onnxruntime>=1.23.2",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "app.py"]

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
cammana = "app:main"


